migrate to flat layout
add pyproject.toml for pip(x) install support

if installed as HA Addon handle options.json
