+++ evseMQTT.service
@@ -1,6 +1,6 @@
 [Unit]
-Description=Communicating with EVSE-based Electric Vehicle Charging Wallboxes using Bluetooth Low Energy (BLE)
-After=network.target
+Description=Communicating with BESEN-based Electric Vehicle Charging Wallboxes using Bluetooth Low Energy (BLE)
+After=network.target dbus.service
 
 [Service]
 ExecStart=/usr/bin/python3 /path/to/evseMQTT/main.py --address <address> --password <password> --unit <unit> --mqtt --mqtt_broker <mqtt_broker> --mqtt_port <mqtt_port>
