--- requirements.txt
@@ -1,2 +1,0 @@
-bleak==0.20.2
-paho-mqtt==1.6.1
